add text padding and text align params for oktoast
